🐛(search) fix header and home search component
Fix the search component that is included on header and on the homepage.
The search was redirecting to the /en/course/ page.
That is configured has a DjangoCMS redirect page.
The redirect lost the `query` URL parameter.
The solution is to set the `courses-search` on the /en/courses page.
Then the search component generate URLs directly to the correct page.
Fixes GN-754