💄(course_detail) higher enroll button
Move enroll button for open to enroll course runs to be higher on the
screen. For a single open course run the enroll button is visible on
the right using a desktop viewport. If using mobile at
least it will be visible for simple swipe instead of scrolling
to the button of the screen.