🐛(test) fix run local tests
After the Richie upgrade, with its Django upgrade, etc.
it was no longer possible to run the local backend tests with the
Makefile target `make test-back`.
This fixes it, making the project and its tests to use the `root` MySQL
database user, instead of using the an application user. So when the
tests are starting up, it can create the new test database on the fly.

fixes fccn/nau-richie-site-factory#250